+++ src/components/ContributionBreakdown.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { get } from 'lodash';
 
 import StyledCard from './StyledCard';
 import { Span } from './Text';
